Update en.lua
This is the English file I use in the 4.1 version of the game. It should be compatible with all small updates made after 4.1.

I changed the versioning to match the game version.

I am planning to re-implement the credits roll with the upcoming update. Your credits for translating the game will be kept in a different file, so do not include them here.